refactor: migrate ConfiguredAuthorization from OC to CSL
Deletes OC's duplicate ConfiguredAuthorization class and redirects all
8 callers to io.camunda.security.api.model.config.initialization.ConfiguredAuthorization.

CSL's version is functionally identical. This change is safe because
AuthorizationConfigurer (the engine-layer consumer) uses ConfiguredAuthorization
only as a config DTO input — it is not a protocol or log-stream type.

Part of camunda-security-library#183 (inc4b).
Depends on: #53388
