+++ qa/acceptance-tests/src/test/java/io/camunda/it/identity/InitializeInvalidAuthorizationIT.java
@@ -11,7 +11,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.camunda.security.api.model.authz.AuthorizationOwnerType;
 import io.camunda.security.api.model.authz.PermissionType;
-import io.camunda.security.configuration.ConfiguredAuthorization;
+import io.camunda.security.api.model.config.initialization.ConfiguredAuthorization;
 import io.camunda.security.validation.IdentityInitializationException;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
