+++ qa/util/src/main/java/io/camunda/qa/util/multidb/EntityManager.java
@@ -25,9 +25,9 @@ import io.camunda.client.api.search.response.RoleUser;
 import io.camunda.client.api.search.response.Tenant;
 import io.camunda.client.api.search.response.User;
 import io.camunda.qa.util.multidb.TestEntityConfigurer.ConfigurationTestEntities;
+import io.camunda.security.api.model.config.initialization.ConfiguredAuthorization;
 import io.camunda.security.api.model.config.initialization.ConfiguredGroup;
 import io.camunda.security.api.model.config.initialization.ConfiguredRole;
-import io.camunda.security.configuration.ConfiguredAuthorization;
 import io.camunda.zeebe.protocol.record.value.EntityType;
 import java.util.HashSet;
 import java.util.List;
