+++ qa/util/src/main/java/io/camunda/qa/util/multidb/TestEntityConfigurer.java
@@ -21,12 +21,12 @@ import io.camunda.security.api.model.authz.AuthorizationOwnerType;
 import io.camunda.security.api.model.authz.AuthorizationResourceType;
 import io.camunda.security.api.model.authz.EntityType;
 import io.camunda.security.api.model.authz.PermissionType;
+import io.camunda.security.api.model.config.initialization.ConfiguredAuthorization;
 import io.camunda.security.api.model.config.initialization.ConfiguredGroup;
 import io.camunda.security.api.model.config.initialization.ConfiguredMappingRule;
 import io.camunda.security.api.model.config.initialization.ConfiguredRole;
 import io.camunda.security.api.model.config.initialization.ConfiguredTenant;
 import io.camunda.security.api.model.config.initialization.ConfiguredUser;
-import io.camunda.security.configuration.ConfiguredAuthorization;
 import java.util.List;
 import java.util.Set;
 import java.util.function.Function;
